+++ Diagnostics/Assets/Basic/LDL/LDLController.cs
@@ -16,6 +16,7 @@ using KLib.Signals;
 using LDL;
 
 using BasicMeasurements;
+using KLib.Wave;
 
 public class LDLController : MonoBehaviour, IRemoteControllable
 {
@@ -60,6 +61,13 @@ public class LDLController : MonoBehaviour, IRemoteControllable
 
         _sliderPanel = _workPanel.GetComponent<LDLSliderPanel>();
         _sliderPanel.LockInPressed += OnLockIn;
+
+        Application.logMessageReceived += HandleException;
+    }
+
+    private void OnDestroy()
+    {
+        Application.logMessageReceived -= HandleException;
     }
 
     private void Start()
@@ -543,4 +551,38 @@ public class LDLController : MonoBehaviour, IRemoteControllable
 
         return s;
     }
+
+    public void HandleException(string condition, string stackTrace, LogType type)
+    {
+        if (type == LogType.Log || type == LogType.Warning)
+        {
+            return;
+        }
+
+        try
+        {
+            _workPanel.SetActive(false);
+        }
+        catch { }
+
+        HandleError(condition, stackTrace);
+    }
+
+    void HandleError(string error, string stackTrace = "")
+    {
+        if (error.Equals("Exception"))
+        {
+            error = "An exception occurred";
+        }
+
+        HTS_Server.SendMessage(_mySceneName, $"Error:{error}");
+        Debug.Log($"[{_mySceneName} error]: {error}{Environment.NewLine}{stackTrace}");
+
+        if (!_isRemote)
+        {
+            ShowFinishPanel("The run was stopped because of an error");
+        }
+    }
 }
+
+