+++ Diagnostics/Assets/Basic/LDL/LDLSliderPanel.cs
@@ -30,10 +30,6 @@ public class LDLSliderPanel : MonoBehaviour
         _textSizeAnimator = _prompt.gameObject.GetComponent<TextSizeAnimator>();
     }
 
-	void Start () 
-    {
-	}
-
     public void Initialize(LDLMeasurementSettings settings, Channel ch)
     {
         _lockInButton.gameObject.SetActive(false);
@@ -46,7 +42,7 @@ public class LDLSliderPanel : MonoBehaviour
 
         for (int k=0; k< _sliders.Count; k++) 
         {
-            _sliders[k].InitializeStimulusGeneration(ch.Clone(), settings.MinLevel, settings.ModDepth_pct);
+            _sliders[k].InitializeStimulusGeneration(ch.Clone(), settings.Bandwidth, settings.MinLevel, settings.ModDepth_pct);
             _sliders[k].SliderMoved += OnSliderMoved;
 
             var rt = _sliders[k].gameObject.GetComponent<RectTransform>();